requirements.txt
# Web框架
fastapi>=0.101.1
uvicorn>=0.23.2
pydantic>=2.3.0
python-multipart==0.0.6

# HTTP请求
cloudscraper==1.2.71
aiohttp>=3.8.5
aiofiles==0.8.0

# 工具
python-dotenv>=1.0.0
requests==2.30.0
pillow==10.0.0
cryptography==41.0.3
psutil==5.9.5
tzdata>=2023.3
pytz>=2023.3
retry>=0.9.2
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.1
tzlocal>=5.0.1
humanize>=4.7.0
prettytable>=3.8.0
natsort>=8.4.0
tenacity>=8.2.3
PyJWT>=2.6.0 